Require unlock for custom SMS dialog
Currently when a call is recieved there is an option to write a custom
text message in response to the call. This could allow an attacker to
imporsonate the user. With this change devices that have a lock screen
set will require that the pattern/pin/password be entered before a
custom message can be written.

This is a cherry-pick from the Pixel dialer. The original commit can be
found at cl/257702865

Bug: b/137102479
Test: Call phone running the modified dialer and attempt to send a
custom message

Change-Id: Ib6822436bcebc799e7e920f1a5898d107dd619db
(cherry picked from commit fab2f0fddbf28cf3ed8608ab9053b2c985d241f7)
CRs-Fixed: 2827003