Fix the contact name can't be updated in time
In some situations, Contacts Provider needs to take more time to
update one contact. If it isn't updated before CalllogFragment gets
resumed, will result in delay of refreshing UI, in the worse case,
the delay should be 1 minute.

To fix it, refresh data directly if the fragment is visible.

Change-Id: I8a80d00a9bb6d331b255b09e56eab19459b32ff0
CRs-Fixed: 2934160