IMS: Fix extra calls created in UI when pulling external call
When external call changes to internal after call pull, it is
removed from external calllist and added to normal calllist.
During this transition, if multiple call details change events
are received at Dialer, there is a chance that the onDetailsChanged
callback is invoked even after the call is removed from the list.
This is because the onDetailsChanged callback is invoked by posting
a Runnable to handler. This causes onCallAdded to be invoked
multiple times for the same external call and duplicate calls get
added to CallList.
To avoid this, check if the call is still tracked by ExternalCallList
before handling the onDetailsChanged callback.
Also remove the PROPERTY_IS_EXTERNAL_CALL assertion during
ExternalCallList.onCallAdded. This is to prevent crash when external
call property is removed immediately after onDetailsChanged Runnable
is posted to handler (refer Bug# 206064937).

Change-Id: Iae3255ff59856a0d5ff77d06e8a1b5da37b2d699
CRs-Fixed: 3155261