Fix issue of call end dialog not displayed sometimes
If a call is added to Dialer CallList in disconnected state
and InCall UI is already started, there is no further trigger
to notify call disconnect to listeners. So call end dialog
is not displayed when MO call is immediately ended for cases
like FDN blocked. Explicitly notify call disconnect in such
cases when call is added to CallList so that call end dialog
is displayed.

Change-Id: I24687073b2a7f9a339c11d9326f62b0a19be1399
CRs-Fixed: 3764866